Factored out ExifDataSaver
