+++ app/src/main/java/ai/qed/camera/CaptureMultipleImagesActivity.kt
@@ -38,7 +38,6 @@ import androidx.camera.view.PreviewView
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import androidx.core.content.FileProvider
-import androidx.exifinterface.media.ExifInterface
 import com.google.android.gms.location.LocationCallback
 import com.google.android.gms.location.LocationRequest
 import com.google.android.gms.location.LocationResult
@@ -56,8 +55,6 @@ import java.util.concurrent.Executor
 import java.util.zip.Deflater
 import java.util.zip.ZipEntry
 import java.util.zip.ZipOutputStream
-import kotlin.math.abs
-import kotlin.math.floor
 
 class CaptureMultipleImagesActivity : AppCompatActivity() {
     private lateinit var imageCapture: ImageCapture
@@ -507,7 +504,7 @@ class CaptureMultipleImagesActivity : AppCompatActivity() {
     private fun saveGeoExifData(photoFile: File) {
         getLocation { location ->
             if (location != null) {
-                addGeoExifData(photoFile, location)
+                ExifDataSaver.saveLocationAttributes(photoFile, location, currentAzimuth, currentPitch, currentRoll)
             }
         }
     }
@@ -538,26 +535,6 @@ class CaptureMultipleImagesActivity : AppCompatActivity() {
             )
         } else {
             onLocationFetched(null)
-        }
-    }
-
-    private fun addGeoExifData(photoFile: File, location: Location) {
-        try {
-            val exif = ExifInterface(photoFile.absolutePath)
-
-            val latitude = location.latitude
-            val longitude = location.longitude
-            exif.setAttribute(ExifInterface.TAG_GPS_LATITUDE, convertToExifFormat(latitude))
-            exif.setAttribute(ExifInterface.TAG_GPS_LATITUDE_REF, if (latitude >= 0) "N" else "S")
-            exif.setAttribute(ExifInterface.TAG_GPS_LONGITUDE, convertToExifFormat(longitude))
-            exif.setAttribute(ExifInterface.TAG_GPS_LONGITUDE_REF, if (longitude >= 0) "E" else "W")
-
-            val gyroData = "Azimuth: $currentAzimuth, Pitch: $currentPitch, Roll: $currentRoll"
-            exif.setAttribute(ExifInterface.TAG_USER_COMMENT, gyroData)
-
-            exif.saveAttributes()
-        } catch (e: IOException) {
-            Log.e("CaptureMultiplePhotos", "Failed to save EXIF data", e)
         }
     }
 
@@ -787,13 +764,5 @@ class CaptureMultipleImagesActivity : AppCompatActivity() {
 
     private fun getExecutor(): Executor {
         return ContextCompat.getMainExecutor(this)
-    }
-
-    private fun convertToExifFormat(coordinate: Double): String {
-        val absolute = abs(coordinate)
-        val degrees = floor(absolute).toInt()
-        val minutes = floor((absolute - degrees) * 60).toInt()
-        val seconds = ((absolute - degrees - minutes / 60.0) * 3600 * 1000).toInt()
-        return "$degrees/1,$minutes/1,$seconds/1000"
     }
 }
